pyproject.toml
[project]
name = "webweaver"
version = "0.1.0"
description = "WebWeaver dual-agent framework for open-ended deep research"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "WebWeaver1" }]
license = { text = "MIT" }

dependencies = [
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.0",
  "readability-lxml>=0.8.1",
  "lxml>=5.2.0",
  "python-dotenv>=1.0.1",
  "openai>=1.40.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "redis>=5.0.0",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "duckduckgo-search>=6.2.13",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.6.0",
]

[project.scripts]
webweaver = "webweaver.cli:app"
webweaver-api = "webweaver.api.serve:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/webweaver"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
